# app/openapi.yml
openapi: 3.0.3
info:
  title: Koppen API voor landelijke vakantie verhuur
  description: Dit is het koppelstuk tussen mijn amsterdam en landelijke vakantie verhuur
  version: 1.0.0
paths:
  /status/health:
    get:
      summary: returns "OK"
      responses:
        "200":
          description: Returns "OK"
          content:
            application/json:
              schema:
                type: string
                enum:
                  - "OK"
  /vakantie-verhuur/get:
    parameters:
      - name: Authorization
        in: header
        description: Bearer token
        required: true
        schema:
          type: string
    get:
      summary: LVV informatie op basis van BSN
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    $ref: "#/components/schemas/LVVInfo"
                  status:
                    type: string
                    enum:
                      - "OK"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum:
            - "ERROR"
        message:
          type: string
    LVVInfo:
      type: array
      items:
        type: object
        properties:
          registrationNumber:
            type: string
            example: AAAA AAAA AAAA AAAA AAAA
          city:
            type: string
            example: Amsterdam
          houseLetter:
            type: string
            nullable: true
            example: a
          houseNumber:
            type: string
            example: 1,
          houseNumberExtension:
            type: string
            nullable: true
            example: 2,
          postalCode:
            type: string
            example: 1012PN
          shortName:
            type: string
            example: Amstel
          street:
            type: string
            example: Amstel
         